+++ libs/src/ui/module/dropdown/index.ts
@@ -1,0 +1,4 @@
+// @index(['./*.tsx', '!./*.stories.tsx'], f => `export * from '${f.path}'`)
+export * from './dropdown';
+export * from './styled';
+// @endindex
